Remove keywordize-keys. Fixes private-message
keywordize-keys was added with the intention of replacing underscores
with dashes. Like xmpp_jid -> xmpp-jid. But it just replaces strings with
keywords. Our map has keywords already, so this was doing nothing, except of
breaking private-message because we assumed xmpp_jid would be xmpp-jid

We might want to use something like https://github.com/jeremyheiler/wharf to
dasherize keyworkds with underscores.